renames mylib to lib
