+++ lib/ApiNavigateableObserver.js
