+++ lib/BlogObserver.js
