+++ lib/NavigatableObserver.js
