+++ lib/StateObserver.js
