+++ lib/StateSubject.js
